Dev (#18)
* Update default settings of the INS demo

* collect all matplotlib-realted functions to sim_data_plot.py

* remove demo that is out of maintenance
