+++ demo_algorithms/dmu380_sim_lib/ekfSim_ins.cfg
@@ -29,9 +29,9 @@ freeIntegrate = false
 useMag = false
 useGps = true
 useOdo = false
-stationaryLockYaw = false
+stationaryLockYaw = true
 dynamicMotion = true
-enableImuStaticDetect = false
+enableImuStaticDetect = true
 // Set the turn-switch threshold
 turnSwitchThreshold = 6.0
 // Set the hard/soft-iron parameters
@@ -41,19 +41,19 @@ softIronScaleRatio = 1.0
 softIronAngle = 0.0
 // Set the acceleration switch parameters (deviation from 1 [g] and period
 //   required to trigger high-gain mode (defaults 0.0055 [g] and 4.5 [sec])
-linAcelSwitchLimit = 0.0055
-linAccelSwitchDelay = 4.0
+linAcelSwitchLimit = 0.012
+linAccelSwitchDelay = 2.0
 // the following is in unit of sec
 freeIntegrationTime = 120.0
 // State duration times (defaults: 0.36, 0.74, 60.0, 30.0)
 stabSysDur = 0.36
 initAttDur = 0.64
-highGainDur = 10.0
-lowGainDur = 1.0
+highGainDur = 30.0
+lowGainDur = 30.0
 // default 3.0 seconds, the following are in units of sec
 maxGpsDropTime = 300.0
 maxReliableDRTime = 10.0
 // Select the LPF type applied to the data before the measurement is converted
 //   roll/pitch angles (0, 5, 10, 25 Hz); default = 10 Hz
-accelLPFType = 10
+accelLPFType = 5
