--- demo_algorithms/dmu380_sim_lib/ekfSim_tilt.cfg
@@ -1,52 +1,0 @@
-// Configuration file for the INS380_Offline simulation
-// 
-// Specify the input and output files, the input data rate, and the output packet type
-inFile = F:\vbShared\dmu380_sim_src_20180604\test_data\noisy_sensor.dat
-inFileDR = 200
-outFile = F:\vbShared\dmu380_sim_src_20180604\test_data\noisy_sensor_results.dat
-outFileDR = 25
-outputPkt = EKF_STATES
-// Specify the DMU version and the rate-sensor type
-dmuVersion = AHRS380ZA-400
-rsType = BMI
-// Specify the product configuration
-hasMags = false
-hasGps = false
-// Specify the operational configuration
-freeIntegrate = false
-useMags = false
-useGps = false
-stationaryLockYaw = false
-dynamicMotion = true
-// Set the turn-switch threshold
-turnSwitchThreshold = 0.5
-// Set the hard/soft-iron parameters
-hardIron_X = 0.0
-hardIron_Y = 0.0
-softIronScaleRatio = 1.0
-softIronAngle = 0.0
-// Set the acceleration switch parameters (deviation from 1 [g] and period
-//   required to trigger high-gain mode (defaults 0.0055 [g] and 4.5 [sec])
-linAcelSwitchLimit = 0.0055
-linAccelSwitchDelay = 4.5
-//
-freeIntegrationTime = 120.0
-// State duration times (defaults: 0.36, 0.74, 60.0, 30.0)
-stabSysDur = 0.36
-initAttDur = 0.74
-highGainDur = 60.0
-lowGainDur = 30.0
-// default 3.0 seconds
-maxGpsDropTime = 3.0
-// Select the LPF type applied to the data before the measurement is converted
-//   roll/pitch angles (0, 5, 10, 25 Hz); default = 10 Hz
-accelLPFType = 10
-//
-procCovarMult_rateBias = 1.0
-procCovarMult_attitude = 1.0
-//
-// For the measurement multipliers, the same applies (the measurement is
-//   applied to the nominal algorithm value of 400.0).
-measCovarMult_roll = 400.0
-measCovarMult_pitch = 400.0
-//
